+++ Sources/Kommunicate/Classes/Views/FeedbackRatingView.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import UIKit
+import ApplozicSwift
 
 class FeedbackRatingView: UIView {
 
@@ -72,6 +73,18 @@ class FeedbackRatingView: UIView {
     private func setupView() {
         onRatingTap = { [weak self] tag in
             self?.selectedRatingTag = tag
+//
+            switch tag {
+            case 1:
+                KMCustomEventHandler.shared.publish(triggeredEvent: CustomEvent.rateConversationEmotionsClick, data: ["UserSelection":["Emoji Clicked": KMCustomEventHandler.CSATRating.poor]])
+            case 5:
+                KMCustomEventHandler.shared.publish(triggeredEvent: CustomEvent.rateConversationEmotionsClick, data: ["UserSelection":["Emoji Clicked": KMCustomEventHandler.CSATRating.average]])
+            case 10:
+                KMCustomEventHandler.shared.publish(triggeredEvent: CustomEvent.rateConversationEmotionsClick, data: ["UserSelection":["Emoji Clicked": KMCustomEventHandler.CSATRating.great]])
+            default:
+                print("Emoji value is not registered!!")
+            }
+
         }
         ratingButtons.forEach { $0.ratingTapped = onRatingTap }
     }